requirements.txt
jupyter
jupyter_kernel_gateway
flask-restful
flask-apidoc

